+++ app/policies/draw_student_policy.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+# Policies for updating students
+class DrawStudentPolicy < ApplicationPolicy
+  def edit?
+    user.admin?
+  end
+
+  def update?
+    edit?
+  end
+
+  def bulk_assign?
+    user.admin?
+  end
+end